Change Noah-MP evaporation code for diag purposes
The surface resistance to evaporation depends on
the depth of the surface layer. By default, this is
set to 0.1, but the user can change the surface layer,
which will in turn change the evaporation from bare
soil. The edit in this commit hard-codes the resistance
parameterization to use 0.1 meters as the surface
layer depth, regardless of what the actual surface
layer depth has been set to in the lis.config file.